Extract nav graphs and simplify parent component
